requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.3.0
django==1.11.16
enum34==1.1.6
ipython-genutils==0.2.0
ipython==5.8.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycrypto==2.6.1
pygments==2.3.0
pytz==2018.7
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
